+++ app/src/components/PianoRoll/PianoRollCanvas/PianoRollCanvas.tsx
@@ -8,7 +8,7 @@ import { useStores } from "../../../hooks/useStores"
 import { Beats } from "../../GLNodes/Beats"
 import { Cursor } from "../../GLNodes/Cursor"
 import { Selection } from "../../GLNodes/Selection"
-import { useNoteMouseGesture } from "../MouseHandler/NoteMouseHandler"
+import { useNoteMouseGesture } from "../MouseHandler/useNoteMouseGesture"
 import { PianoRollStageProps } from "../PianoRollStage"
 import { PianoSelectionContextMenu } from "../PianoSelectionContextMenu"
 import { GhostNotes } from "./GhostNotes"
